requirements.txt
asgiref==3.8.1
blinker==1.4
cryptography==3.4.8
Django==4.2.6
django-cors-headers==4.4.0
httplib2==0.20.2
importlib-metadata==4.6.4
PyJWT==2.3.0
PyYAML==6.0
sqlparse==0.5.1
gunicorn==23.0.0
djangorestframework==3.14.0
psycopg2-binary>=2.9
psycopg>=3.0
moviepy





